Add xAlias URL login
