+++ src/pages/Unlock/Unlock.tsx
@@ -11,7 +11,6 @@ import {
   OperaWalletLoginButton,
   WalletConnectLoginButton,
   WebWalletLoginButton as WebWalletUrlLoginButton,
-  XaliasLoginButton,
   XaliasCrossWindowLoginButton,
   CrossWindowLoginButton
 } from 'components/sdkDappComponents';
@@ -19,7 +18,7 @@ import { nativeAuth } from 'config';
 import { RouteNamesEnum } from 'localConstants';
 import { useNavigate } from 'react-router-dom';
 import { AuthRedirectWrapper } from 'wrappers';
-import { WebWalletLoginWrapper, WebWalletLoginConfigEnum } from './components';
+import { WebWalletLoginWrapper, XaliasLoginWrapper } from './components';
 
 type CommonPropsType =
   | OperaWalletLoginButtonPropsType
@@ -79,10 +78,8 @@ export const Unlock = () => {
               customWalletAddress='https://127.0.0.1:3000'
               {...commonProps}
             />
-            <WebWalletLoginWrapper
-              {...commonProps}
-              config={['crossWindow', 'url']}
-            />
+            <XaliasLoginWrapper {...commonProps} />
+            <WebWalletLoginWrapper {...commonProps} />
           </div>
         </div>
       </div>
