+++ src/pages/Unlock/components/index.tsx
@@ -1,1 +1,2 @@
 export * from './WebWalletLoginWrapper';
+export * from './XaliasLoginWrapper';
